+++ src/classes/UI5Classes/UI5Parser/UIClass/JSClass.ts
@@ -1,5 +1,4 @@
-import { AbstractUIClass, UIMethod } from "./AbstractUIClass";
-import { MainLooper } from "../../JSParser/MainLooper";
+import { AbstractUIClass, UIField, UIMethod } from "./AbstractUIClass";
 function getAllFuncs(toCheck: any) {
 	let props: string[] = [];
 	let obj = toCheck;
@@ -11,7 +10,7 @@ function getAllFuncs(toCheck: any) {
 		if (e!=arr[i+1] && typeof toCheck[e] == 'function') return true;
 	});
 }
-const classData: {[key: string]: {methods: UIMethod[]}} = {
+const classData: {[key: string]: {methods: UIMethod[], fields: UIField[]}} = {
 	Promise: {
 		methods: [{
 			name: "then",
@@ -33,7 +32,8 @@ const classData: {[key: string]: {methods: UIMethod[]}} = {
 			description: "Promise .finally",
 			returnType: "Promise",
 			visibility: "public"
-		}]
+		}],
+		fields: []
 	},
 	array: {
 		methods: getAllFuncs([]).reduce((accumulator: UIMethod[], key: any) => {
@@ -48,7 +48,34 @@ const classData: {[key: string]: {methods: UIMethod[]}} = {
 			}
 
 			return accumulator;
-		}, [])
+		}, []),
+		fields: [{
+			name: "length",
+			description: "Length of an array",
+			type: "number",
+			visibility: "public"
+		}]
+	},
+	Array: {
+		methods: getAllFuncs([]).reduce((accumulator: UIMethod[], key: any) => {
+			if (Array.prototype[key] instanceof Function) {
+				accumulator.push({
+					name: key,
+					params: [],
+					description: key,
+					returnType: "array",
+					visibility: "public"
+				});
+			}
+
+			return accumulator;
+		}, []),
+		fields: [{
+			name: "length",
+			description: "Length of an array",
+			type: "number",
+			visibility: "public"
+		}]
 	},
 	string: {
 		methods: getAllFuncs("").reduce((accumulator: UIMethod[], key: any) => {
@@ -63,7 +90,24 @@ const classData: {[key: string]: {methods: UIMethod[]}} = {
 			}
 
 			return accumulator;
-		}, [])
+		}, []),
+		fields: []
+	},
+	String: {
+		methods: getAllFuncs("").reduce((accumulator: UIMethod[], key: any) => {
+			if (Array.prototype[key] instanceof Function) {
+				accumulator.push({
+					name: key,
+					params: [],
+					description: key,
+					returnType: "string",
+					visibility: "public"
+				});
+			}
+
+			return accumulator;
+		}, []),
+		fields: []
 	}
 };
 export class JSClass extends AbstractUIClass {
@@ -71,5 +115,6 @@ export class JSClass extends AbstractUIClass {
 		super(className);
 
 		this.methods = classData[className].methods;
+		this.fields = classData[className].fields;
 	}
 }