+++ src/classes/filerenaming/handlers/abstraction/FileRenameHandler.ts
@@ -6,7 +6,6 @@ import * as path from "path";
 import { UIClassFactory } from "../../../UI5Classes/UIClassFactory";
 const fileSeparator = path.sep;
 
-
 export abstract class FileRenameHandler {
 	abstract handleFileRename(oldUri: vscode.Uri, newUri: vscode.Uri) : void;
 
@@ -34,22 +33,22 @@ export abstract class FileRenameHandler {
 		for (const wsFolder of wsFolders) {
 			const workspaceFilePaths = glob.sync(wsFolder.uri.fsPath.replace(/\\/g, "/") + "/" + src + "/**/*{.js,.xml,.json}");
 			workspaceFilePaths.forEach(filePath => {
-				let file = fs.readFileSync(filePath, "utf8");
-				if (file.indexOf(textToReplaceFromDotNotation) > -1 || file.indexOf(textToReplaceFromSlashNotation) > -1) {
-					file = file.replace(new RegExp('\\"' + textToReplaceFromDotNotation.replace(/\./g, "\\.") + '\\"', "g"), '"' + textToReplaceToDotNotation + '"');
-					file = file.replace(new RegExp('\\"' + textToReplaceFromSlashNotation.replace(/\./g, "\\.") + '\\"', "g"), '"' + textToReplaceToSlashNotation + '"');
+				let fileContent = fs.readFileSync(filePath, "utf8");
+				if (fileContent.indexOf(textToReplaceFromDotNotation) > -1 || fileContent.indexOf(textToReplaceFromSlashNotation) > -1) {
+					fileContent = fileContent.replace(new RegExp('\\"' + textToReplaceFromDotNotation.replace(/\./g, "\\.") + '\\"', "g"), '"' + textToReplaceToDotNotation + '"');
+					fileContent = fileContent.replace(new RegExp('\\"' + textToReplaceFromSlashNotation.replace(/\./g, "\\.") + '\\"', "g"), '"' + textToReplaceToSlashNotation + '"');
 					//TODO: Think how to do it async. Sync currently needed for folder rename, where mass file change is fired and
 					//there might be multiple changes for the same file
-					fs.writeFileSync(filePath, file);
+					fs.writeFileSync(filePath, fileContent);
 
 					//TODO: Use observer pattern here
 					if (filePath.endsWith(".js")) {
 						const classNameOfTheReplacedFile = FileReader.getClassNameFromPath(filePath.replace(/\//g, fileSeparator));
 						if (classNameOfTheReplacedFile) {
-							UIClassFactory.setNewCodeForClass(classNameOfTheReplacedFile, file);
+							UIClassFactory.setNewCodeForClass(classNameOfTheReplacedFile, fileContent);
 						}
 					} else if (filePath.endsWith(".view.xml")) {
-						FileReader.setNewViewContentToCache(file, filePath);
+						FileReader.setNewViewContentToCache(fileContent, filePath);
 					}
 				}
 			});