+++ src/classes/xmllinter/parts/UnusedNamespaceLinter.ts
@@ -7,7 +7,7 @@ export class UnusedNamespaceLinter extends Linter {
 
 		const aPrefixes = document.match(/(?<=xmlns:).*?(?==)/g);
 		aPrefixes?.forEach(prefix => {
-			const aPrefixes = new RegExp(`(<|\\s)${prefix}:`, "g").exec(document);
+			const aPrefixes = new RegExp(`(<|\\s)${prefix.trim()}:`, "g").exec(document);
 			if (!aPrefixes || aPrefixes.length === 0) {
 				const positionBegin = document.indexOf(`xmlns:${prefix}=`);
 				const position = LineColumn(document).fromIndex(positionBegin - 1);